e2e tests: reorganize analytics specs across serial/parallel pools (TESTOPS-210)
Re-balance the four analytics E2E specs between core-serial and
core-parallel so more of them run in parallel, based on the two hazards
they share: toggling the global woocommerce_analytics_scheduled_import
option, and asserting store-wide totals.

- Renamed analytics-data.spec.ts -> analytics.spec.ts and moved it to
  core-parallel. Its assertions are already product-scoped (#66371). The
  Overview "manual import trigger" tests (a mode toggler) are folded in as
  a describe so both toggle the import mode within one worker instead of
  racing across workers; the scheduled-mode test drains the Action
  Scheduler queue (?process-waiting-actions) so the "now" batch import
  clears and the trigger button settles to idle under wp-env's no-cron.
- analytics-overview -> core-parallel. Now layout-only (the toggler block
  moved to analytics.spec.ts); it mutates only the admin's own
  dashboard_sections meta.
- analytics-settings stays serial: as a standalone file it would race
  analytics.spec.ts over the import-mode option in the parallel pool.
- analytics-access deleted: a thin @not-e2e smoke test of the Home
  Stats-overview deep-links, redundant with the reports the other specs
  already load.

Validated: 10x core-parallel runs, analytics specs green in all 10.
