+++ plugins/woocommerce/tests/e2e/playwright.config.ts
@@ -120,19 +120,17 @@ export const setupProjects = [
  * `core-parallel` by default, except the other-project folders in `nonCoreSpecs`.
  */
 const serialRunSpecs = [
-	// Forces immediate analytics-import mode by flipping the global
-	// `woocommerce_analytics_scheduled_import` option, and needs it to stay
-	// immediate while it creates orders and drains their imports. In
-	// `core-parallel` that would race `analytics-settings.spec.ts` — which runs in
-	// that pool and toggles the same option — over the shared store and Action
-	// Scheduler queue. Its numeric assertions are scoped to its own products, so
-	// they no longer depend on store-wide totals.
-	'**/tests/analytics/analytics-data.spec.ts',
-	// Asserts store-wide `$0.00 / Orders 0`, polluted by concurrent orders.
-	'**/tests/analytics/analytics-access.spec.ts',
-	// Mutates the shared admin's `woocommerce_meta.dashboard_sections` and flips the
-	// global `woocommerce_analytics_scheduled_import` option (racing analytics-settings).
-	'**/tests/analytics/analytics-overview.spec.ts',
+	// Toggles the global `woocommerce_analytics_scheduled_import` option to
+	// exercise the Settings-page scheduled/immediate switch. Kept serial because,
+	// as a standalone file, running it in `core-parallel` would race
+	// `analytics.spec.ts` — which also toggles that option — across workers.
+	// (The other analytics specs are parallel: `analytics` owns its own
+	// import-mode toggles plus the Overview manual-trigger tests in one file, so
+	// those serialise within a single worker; `analytics-overview` only mutates
+	// the admin's own `dashboard_sections` meta. No other parallel spec depends on
+	// the order-import mode, and this serial job never runs concurrently with the
+	// parallel one.)
+	'**/tests/analytics/analytics-settings.spec.ts',
 	// Flips the global `woocommerce_default_customer_address` (geolocation) and
 	// `woocommerce_enable_ajax_add_to_cart` settings, which change add-to-cart
 	// behavior for every other worker. (`cart.spec.ts` runs in core-parallel — it
