+++ plugins/woocommerce/tests/e2e/tests/analytics/analytics-overview.spec.ts
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import { test, expect, request } from '@playwright/test';
+import { test, expect } from '@playwright/test';
 import type { Page, Locator } from '@playwright/test';
 
 /**
@@ -10,7 +10,6 @@ import type { Page, Locator } from '@playwright/test';
 import { admin } from '../../test-data/data';
 import { tags } from '../../fixtures/fixtures';
 import { ADMIN_STATE_PATH } from '../../playwright.config';
-import { setOption } from '../../utils/options';
 
 const EXPECTED_SECTION_HEADERS = [ 'Performance', 'Charts', 'Leaderboards' ];
 
@@ -300,92 +299,6 @@ test.describe(
 			await test.step( `Expect the Performance section to be added back.`, async () => {
 				await expect( heading_performance ).toBeVisible();
 			} );
-		} );
-	}
-);
-
-test.describe(
-	'Analytics Overview - Manual Import Trigger',
-	{ tag: [ tags.PAYMENTS, tags.SERVICES ] },
-	() => {
-		test.use( { storageState: ADMIN_STATE_PATH } );
-
-		test.beforeAll( async ( { browser } ) => {
-			page = await browser.newPage();
-		} );
-
-		test.beforeEach( async () => {
-			await test.step( `Go to Analytics > Overview`, async () => {
-				await page.goto(
-					'wp-admin/admin.php?page=wc-admin&path=%2Fanalytics%2Foverview'
-				);
-			} );
-		} );
-
-		test.afterAll( async () => {
-			await page.close();
-		} );
-
-		test( 'should show manual update trigger in scheduled mode', async ( {
-			baseURL,
-		} ) => {
-			// Set to scheduled mode
-			await setOption(
-				request,
-				baseURL,
-				'woocommerce_analytics_scheduled_import',
-				'yes'
-			);
-
-			// Reload the page
-			await page.reload();
-
-			// Verify import status bar is visible
-			await expect( page.getByText( 'Data status' ) ).toBeVisible();
-			// Verify "Update now" button is visible
-			const updateButton = page.getByRole( 'button', {
-				name: 'Manually trigger analytics data import',
-			} );
-			await expect( updateButton ).toBeVisible();
-
-			// Click "Update now" button
-			const responsePromise = page.waitForResponse(
-				( response ) =>
-					response
-						.url()
-						.includes( '/wc-analytics/imports/trigger' ) &&
-					response.ok()
-			);
-
-			await updateButton.click();
-
-			// Verify button shows loading state (isBusy)
-			// After clicking, the aria-label changes to "Analytics data import in progress"
-			const busyButton = page.getByRole( 'button', {
-				name: 'Analytics data import in progress',
-			} );
-			await expect( busyButton ).toBeDisabled();
-
-			// Wait for API response
-			await responsePromise;
-		} );
-
-		test( 'should hide manual update trigger in immediate mode', async ( {
-			baseURL,
-		} ) => {
-			// Set to immediate mode
-			await setOption(
-				request,
-				baseURL,
-				'woocommerce_analytics_scheduled_import',
-				'no'
-			);
-
-			// Reload the page
-			await page.reload();
-
-			// Verify import status bar wrapper is NOT rendered
-			await expect( page.getByText( 'Data status' ) ).toBeHidden();
 		} );
 	}
 );
